Created Readme
This application consists of three scenes to demonstrate first person interaction.   The first person interaction is simple:  simply walk into one of the colored cubes and you will be transported to a similarly colored room with the default fading transition of the future room.

The scene-to-transport-to, color and the time-to-fade can be customized by changing the inspector values found in the Portal Behavior (script) component found in the portalCube<color> objects.
Load Scene Number: designates the scene-to-transport-to number.  Acceptable values are 1,2,3
Transition time: the amount of time to fade before transitioning to the next scene
Fade Color: The color of the fade before transporting to the scene

Requirements pasted below instructions:

Software Requirements:
1. Create a Unity project with 3 different simple scenes (they don't have to be fancy - ie. a simple geometric plane with a square game object)

2. Download the Unity Character FPS asset.

3. Write some simple scripts in C# to do the following: When the FPS character walks into a game object, the scene fades out and changes to the next scene, which fades in.

4. Please include the following:

    A. A way for the scene number (that the scene switches to) to be selected in the inspector GUI.

    B. A way to set the color and duration of the fade in/out.